refactor Sam to own service, other various cleanup
